Modification compte user

+++ Back-End/controller/user.js
@@ -11,7 +11,8 @@ module.exports = {
     register: function(req, res){
         console.log("==> POST REGISTER");
         var donnee = req.body;
-        var email = donnee.email, mdp1 = donnee.pass, mdp2 = donnee.re_pass;
+        console.log(donnee);
+        var email = donnee.email, mdp1 = donnee.pass, mdp2 = donnee.re_pass, nom = donnee.nom, prenom = donnee.prenom, ville = donnee.ville, adresse = donnee.ville, pdp = "";
         connexion.then(function(db){
             service.inscrire(email, mdp1, mdp2, db).then(function(verification){
                 if(verification !== true){
@@ -20,7 +21,10 @@ module.exports = {
                 else{
                     bcrypt.hash(mdp1, saltRounds, function(err, hash) {
                         if(err) throw(err);
-                        db.query("INSERT INTO user(email, mdp) VALUES(?,?)", [email, hash], function(err, resultat){
+                        if(req.files){              
+                            pdp = service.uploadfile("image/", req.files.pdp);
+                        } 
+                        db.query("INSERT INTO user(email, mdp, nom, prenom, ville, adresse, pdp) VALUES(?,?,?,?,?,?,?)", [email, hash, nom, prenom, ville, adresse, pdp], function(err, resultat){
                             if(err) return res.status(500).send("Erreur: ressource");
                             service.return_user_email(email, db).then(function(resultat){
                                 if(resultat == "Vérifier votre adresse email"){
@@ -69,17 +73,27 @@ module.exports = {
 
     update: function(req, res){
         console.log("==> UPDATE USER");
-        var id = req.body.id, nom = req.body.nom, prenom = req.body.prenom, pdp = "", video = "";
+        console.log(req.body);
+        var id = req.body.id, email = req.body.email, nom = req.body.nom, prenom = req.body.prenom, adresse = 
+        req.body.adresse, ville = req.body.ville, pdp = "";
         connexion.then(function(db){
             service.verifier_exist_user(id, db).then(function(user){
-                if(user){
-                    if(req.files){
-                        if(user[0].pdp) service.deleteFile("public/"+user[0].pdp);                
-                        pdp = service.uploadfile("image/", req.files.pdp);
-                    }       
-                    db.query("UPDATE user set nom = ?, prenom = ?, pdp = ?, video = ? WHERE id =?", [nom, prenom, pdp, video, id], function(err){
-                        if(err) return res.status(500).send({error: "Ressource"});
-                        res.send({success:true});
+                if(user && email){
+                    service.mail_utiliser_autre_user(email, id, db).then(function(verification){
+                        console.log(verification);
+                        if(verification === false){
+                            if(req.files){
+                                if(user[0].pdp) service.deleteFile("public/"+user[0].pdp);                
+                                pdp = service.uploadfile("image/", req.files.pdp);
+                            }       
+                            db.query("UPDATE user set email = ?, nom = ?, prenom = ?, pdp = ?, ville = ?, adresse = ? WHERE id =?", [email, nom, prenom, pdp, ville, adresse, id], function(err){
+                                if(err) return res.status(500).send({error: "Ressource"});
+                                res.send({success:true});
+                            })
+                        }
+                        else{
+                            res.status(403).send({error:verification});
+                        }
                     })
                 }
                 else{
